[ refactor ] gridValue core파일로 이동
