+++ src/components/Category/CategoryContents/index.tsx
@@ -1,5 +1,5 @@
 import { useCategoryLists } from "../../../core/api/main";
-import { categoryList } from "../../../core/category/categoryList";
+import { categoryList, gridValue } from "../../../core/category/categoryList";
 import { St } from "./style";
 
 export default function CategoryContents() {
@@ -7,78 +7,6 @@ export default function CategoryContents() {
   // 이걸로 하면돼!!!!! import 받아온 거 파일 없애주고,
 
   // 아래 gridValue는 core 폴더에 넣어주면 좋을 것 같아요!~
-  type gridValueProps = {
-    id: number;
-    columnStart: number;
-    columnEnd: number;
-    rowStart: number;
-    rowEnd: number;
-  };
-  const gridValue: gridValueProps[] = [
-    {
-      id: 1,
-      columnStart: 1,
-      columnEnd: 2,
-      rowStart: 1,
-      rowEnd: 6,
-    },
-
-    {
-      id: 2,
-      columnStart: 1,
-      columnEnd: 2,
-      rowStart: 6,
-      rowEnd: 9,
-    },
-
-    {
-      id: 3,
-      columnStart: 1,
-      columnEnd: 2,
-      rowStart: 9,
-      rowEnd: 12,
-    },
-
-    {
-      id: 4,
-      columnStart: 1,
-      columnEnd: 2,
-      rowStart: 12,
-      rowEnd: 15,
-    },
-
-    {
-      id: 5,
-      columnStart: 2,
-      columnEnd: 3,
-      rowStart: 1,
-      rowEnd: 4,
-    },
-
-    {
-      id: 6,
-      columnStart: 2,
-      columnEnd: 3,
-      rowStart: 4,
-      rowEnd: 7,
-    },
-
-    {
-      id: 7,
-      columnStart: 2,
-      columnEnd: 3,
-      rowStart: 7,
-      rowEnd: 12,
-    },
-
-    {
-      id: 8,
-      columnStart: 2,
-      columnEnd: 3,
-      rowStart: 12,
-      rowEnd: 15,
-    },
-  ];
 
   return (
     <St.FlexContainer>
