+++ src/core/category/categoryList.ts
@@ -46,3 +46,76 @@ export const categoryList: categoryListProps[] = [
     content: "커플과의 추억을 공유하세요",
   },
 ];
+
+type gridValueProps = {
+  id: number;
+  columnStart: number;
+  columnEnd: number;
+  rowStart: number;
+  rowEnd: number;
+};
+export const gridValue: gridValueProps[] = [
+  {
+    id: 1,
+    columnStart: 1,
+    columnEnd: 2,
+    rowStart: 1,
+    rowEnd: 6,
+  },
+
+  {
+    id: 2,
+    columnStart: 1,
+    columnEnd: 2,
+    rowStart: 6,
+    rowEnd: 9,
+  },
+
+  {
+    id: 3,
+    columnStart: 1,
+    columnEnd: 2,
+    rowStart: 9,
+    rowEnd: 12,
+  },
+
+  {
+    id: 4,
+    columnStart: 1,
+    columnEnd: 2,
+    rowStart: 12,
+    rowEnd: 15,
+  },
+
+  {
+    id: 5,
+    columnStart: 2,
+    columnEnd: 3,
+    rowStart: 1,
+    rowEnd: 4,
+  },
+
+  {
+    id: 6,
+    columnStart: 2,
+    columnEnd: 3,
+    rowStart: 4,
+    rowEnd: 7,
+  },
+
+  {
+    id: 7,
+    columnStart: 2,
+    columnEnd: 3,
+    rowStart: 7,
+    rowEnd: 12,
+  },
+
+  {
+    id: 8,
+    columnStart: 2,
+    columnEnd: 3,
+    rowStart: 12,
+    rowEnd: 15,
+  },
+];
